Only render if mountTarget is present
In Jetbuilt, we currently inject CKEditor4 into the initialContent, which for some reason briefly removes the body tag that we also pass.

To workaround this, check to see if the document body is present before returning it as the mountTarget, and skip rendering into the container if the mountTarget is not present.